Refactored BlogListPageController to pull heavy logic out into a controler service in order to support separation of concerns
